Use cosock.asyncify "ssl.https" for everything.
On newer hub firmware versions, non-asyncified sockets are no longer
allowed, and cause all requests to immediately time out.

Further, due to how pulling in the http driver pulls in socket, which
lazily pulls in some ssl pieces, we need to use ssl.https for HTTP
sockets as well.

This does that, and it at least allows https requests to work properly.